refactor Builder class moved
