+++ src/Dacapo/Infra/Adapter/Builder/MysqlDatabaseBuilder.php
@@ -1,10 +1,11 @@
 <?php declare(strict_types=1);
 
-namespace UcanLab\LaravelDacapo\Dacapo\UseCase\Builder;
+namespace UcanLab\LaravelDacapo\Dacapo\Infra\Adapter\Builder;
 
 use UcanLab\LaravelDacapo\Dacapo\Domain\Entity\Schema;
+use UcanLab\LaravelDacapo\Dacapo\UseCase\Shared\Builder\DatabaseBuilder;
 
-class PostgresqlDatabaseBuilder implements DatabaseBuilder
+class MysqlDatabaseBuilder implements DatabaseBuilder
 {
     /**
      * @return bool
@@ -21,7 +22,7 @@ class PostgresqlDatabaseBuilder implements DatabaseBuilder
     public function makeTableComment(Schema $schema): string
     {
         $str = PHP_EOL . PHP_EOL . '        ';
-        $str .= sprintf('DB::statement("COMMENT ON TABLE %s IS \'%s\';");', $schema->getTableName(), $schema->getTableComment());
+        $str .= sprintf('DB::statement("ALTER TABLE %s COMMENT \'%s\'");', $schema->getTableName(), $schema->getTableComment());
 
         return $str;
     }
